+++ tests/fixtures/errors/nomethod.rb
@@ -1,0 +1,9 @@
+def caller
+  victim
+end
+
+def victim
+  nil.foo
+end
+
+caller